functions/stage-embed-articles/requirements.txt
functions-framework==3.*
duckdb==1.3.2
google-cloud-secret-manager
google-genai
pinecone
beautifulsoup4
lxml
langchain==1.0.3
langchain-text-splitters